+++ package.json
@@ -1,0 +1,33 @@
+{
+  "name": "file-manager",
+  "version": "1.0.0",
+  "description": "[![Latest Version on Packagist][ico-version]][link-packagist]\r [![Total Downloads][ico-downloads]][link-downloads]\r [![Build Status][ico-travis]][link-travis]\r [![StyleCI][ico-styleci]][link-styleci]",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gema/laravel-file-manager.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/gema/laravel-file-manager/issues"
+  },
+  "homepage": "https://github.com/gema/laravel-file-manager#readme",
+  "devDependencies": {
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "babel-loader": "^8.2.2",
+    "css-loader": "^5.0.1",
+    "mini-css-extract-plugin": "^1.3.3",
+    "sass": "^1.32.2",
+    "sass-loader": "^10.1.0",
+    "style-loader": "^2.0.0",
+    "webpack": "^5.12.2",
+    "webpack-cli": "^4.3.1"
+  },
+  "dependencies": {}
+}
