+++ src/config/file-manager.php
@@ -1,21 +1,8 @@
 <?php
 
-use App\Models\Visit;
-
 return [
-    'parents' => [
-        Visit::class,
-    ],
+    'parents' => [],
     'filter' => (function($query){
-
-        // Sample for xplora
-
-        // if(admin()){
-        //     return $query
-        //         ->whereIn('parent_id', backpack_user()->visitsIds())
-        //         ->where('parent_type', Visit::class);
-        // }
-
         return $query;
     })
 ];
