+++ src/public/js/index.js
@@ -1,0 +1,3 @@
+(() => {
+    console.log('hello from pckg!')
+})()
