+++ src/resources/js/app.js
@@ -1,0 +1,8 @@
+import '../sass/style.scss';
+
+const {init} = require('./file-manager');
+
+$( document ).ready(function() {
+    console.log('initing filemanager')
+    init();
+});
