+++ src/resources/lang/en/messages.php
@@ -1,5 +1,43 @@
 <?php
 
 return [
-    'test' => 'en',
+    'media_explorer' => 'Media Explorer',
+    'media_tags' => 'Media Tags',
+    'media_tag' => 'Media Tag',
+    'media_types' => 'Media Types',
+    'media_type' => 'Media Type',
+    'media_versions' => 'Media Versions',
+    'media_version' => 'Media Version',
+    'media_manager' => 'Media Manager',
+	'media' => 'Media',
+	'medias' => 'Medias',
+	'title' => 'title',
+	'description' => 'description',
+	'uploading' => 'uploading',
+	'upload' => 'upload',
+	'asign_tag' => 'asign tag',
+	'remove_tag' => 'remove tag',
+	'refresh' => 'resfresh',
+	'close' => 'close',
+	'edit' => 'edit',
+	'save' => 'save',
+	'selected' => 'selected',
+	'no' => 'no',
+	'browse' => 'browse',
+	'parent' => 'parent',
+	'save_selection' => 'save selection',
+    'select_medias_first'=> 'Select some medias first',
+	'tags_unsigned'=> 'Tags removed with success',
+	'tags_unsigned_error' => 'There was an error removing tag',
+	'tags_asigned' => 'Tag asigned with sucees',
+	'tags_asigned_error' => 'There was an error asigning tag',
+	'media_updated' => 'media edited',
+	'no_description' => 'No description asigned',
+	'no_medias_selected' => 'No medias selected',
+	'medias_selected' => 'media selected',
+	'no_medias_found' => 'No media found',
+	'media_type' => 'Media Type',
+	'crop' => 'crop',
+	'image_cropped' => 'image cropped with success',
+	'invalid_metadata' => 'Provided metadata is invalid'
 ];
