+++ src/resources/lang/pt/messages.php
@@ -11,6 +11,8 @@ return [
     'media_manager' => 'Gestor de mídia',
 	'media' => 'mídia',
 	'medias' => 'mídias',
+	'title' => 'titulo',
+	'description' => 'descrição',
 	'uploading' => 'a carregar',
 	'upload' => 'carregar',
 	'asign_tag' => 'atribuir etiqueta',
@@ -24,21 +26,20 @@ return [
 	'browse' => 'procurar',
 	'parent' => 'pai',
 	'save_selection' => 'salvar seleção',
-	
-    // 'Select some medias first'=> 'Selecione algum mídia primeiro',
-	// 'Tags unsigned successfully'=> 'Etiquetas removidas com sucesso',
-	// 'There was an error unsigning tags' => 'Ocorreu um erro ao remover a etiqueta',
-	// 'Tags asigned successfully' => 'Etiquetas atribuidas com sucesso',
-	// 'There was an error asigning tags' => 'Ocorreu um erro ao atribuir a etiqueta',
-	// 'media updated' => 'mídia editado',
-	// 'No description provided' => 'Descrição não atribuida',
-	// 'No medias selected' => 'Sem mídias selecionados',
-	// 'medias selected' => 'mídias selecionados',
-	// 'No medias found' => 'Nenhum mídia encontrado',
-	// 'Media type' => 'Tipo de mídia',
-	// 'crop' => 'cortar',
-	// 'Image cropped successfully' => 'imagem cortada com sucesso',
-	// 'Invalid metadata provided' => 'Os metadados fornecidos são inválidos'
+    'select_medias_first'=> 'Selecione algum mídia primeiro',
+	'tags_unsigned'=> 'Etiquetas removidas com sucesso',
+	'tags_unsigned_error' => 'Ocorreu um erro ao remover a etiqueta',
+	'tags_asigned' => 'Etiquetas atribuidas com sucesso',
+	'tags_asigned_error' => 'Ocorreu um erro ao atribuir a etiqueta',
+	'media_updated' => 'mídia editado',
+	'no_description' => 'Descrição não atribuida',
+	'no_medias_selected' => 'Sem mídias selecionados',
+	'medias_selected' => 'mídias selecionados',
+	'no_medias_found' => 'Nenhum mídia encontrado',
+	'media_type' => 'Tipo de mídia',
+	'crop' => 'cortar',
+	'image_cropped' => 'imagem cortada com sucesso',
+	'invalid_metadata' => 'Os metadados fornecidos são inválidos'
 ];
