+++ src/resources/views/field/file-manager.blade.php
@@ -81,7 +81,6 @@ if(isset($entry)){
         </div>
     </div>
 
-
 {{-- ########################################## --}}
 {{-- Extra CSS and JS for this particular field --}}
 {{-- If a field type is shown multiple times on a form, the CSS and JS will only be loaded once --}}
