+++ src/resources/views/filemanager-page.blade.php
@@ -3,10 +3,10 @@
 @section('content')
 
 <div class="container h-100">
-    <div class="row h-100 justify-content-center align-items-center">
+    <div class="row h-100 justify-content-center">
         <div class="col-sm-12">
             <h1 class="my-4">{{__("file-manager::messages.media_manager")}}</h1>
-            @include("file-manager::filemanager");
+            @include("file-manager::filemanager")
         </div>
     </div>
 </div>
