+++ src/routes/routes.php
@@ -36,6 +36,7 @@ Route::group(
         function () {
             Route::group(['prefix' => 'media'], function () {
                 Route::get('/', [MediaAPIController::class, 'getMedias']);
+                Route::get('/parent', [MediaAPIController::class, 'getParents']);
                 Route::post('/tag', [MediaAPIController::class, 'addTags']);
                 Route::post('/tag/unsign', [MediaAPIController::class, 'removeTags']);
                 Route::post('/upload', [MediaAPIController::class, 'uploadMedia']);
